Setup ASP.NET CORE with React.js and Dapper. Basic User CRUD. Repository pattern.
The project was generated using the 'ASP.NET Core with React.js' template, which was left unmodified as an initial example. In this commit, we incorporated Dapper to implement a basic 'UserController' for performing CRUD operations. We followed the repository pattern for structuring our project. Additionally, we created tables in the phpMySql database, and an Entity-Relationship Diagram (ERD) was included as a new document in the 'Documentation' folder. To maintain consistency, we also updated the names of certain files in the 'Documentation' folder to adhere to a unified naming convention.