Created UserContext, fixed NavMenu, handing JWT  google token
The User Context was created to help share the logged in with google
user variable through out the other components in the application.

The NavMenu is displaying only home and singin when then user is not logged
in.
The JWT token data is stored in context provider in the form of user
variable in order to be shared with other components in the React
front-end.