ControlGate page semi-working
Fixed problems with CORS from the google tocken logic. User is persisted
when the page is refreshed correctly with the image. A lodaing message
is displayed until the server is up. Dapper is configured to ignore
underscore fields in the database. Database tables and fields updated.
Enter parkign button registers the car with inputed license plate and
assocaites it to the currently logged in user if previously not done.
Field validation is not implemented yet.